more generic topmost ordering
There is a problem with many different variations of spelling and ordering of things
that should be "standard" but really are not.

To avoid problems with detecting right order and to avoid unnecessary changes later,
have a more generic handling of figuring out which is the "topmost" entry from a certain
set of possibilities.

Also, more advanced sanity checking should be possible now.
And adding new cases should be simpler as there is much less repetition of the strings.

In summary, code improvements of large scale with many benefits.